Remove unnecessary `clamp`
A route has 2 distinct endpoints, and the minimum cost of any `Tile` is `2`, so the minimum route cost is `4`. Hence this `clamp` isn't needed.

Additionally, the `clamp` was used when some costs were less than `1.0`, so the could would avoid multipling when the division was less than `1.0`. This will never happen.