Remove `oreMovementRemainder` calculation
Remove arbitrarily assignment of excess capacity to the last type of ore. Now that we multiple by the number of trucks before dividing by the route cost, we can make use of fractional capacity by adding more trucks so the fractions add up to a new unit of ore.

This keeps all ore types equal, so there is no odd remainder assignment that needs to be explained. It was particularly odd when each of the other 3 components were all a bit short, and the last type got 3 extra above the other types. It might have made more sense to distribute the excess evenly. Though perhaps better to not have inconsistency between ore type.