Be explicit about making a base class call
The constructor will already set an animation as part of the base class construction, so no need to set it a second time here with the derived class overload. We simply need to set the member field.

Note that we can't use the simpler base class constructor without an explicit animation as that will attempt to set an animation that is invalid for the tube sprite.