Remove no longer necessary check from `addStructure`
Now that we control access to `addStructure`, we can make more assumptions about the parameters it will be passed. In particular, the `Structure` parameter is guaranteed to be a `new` instance, so will never match anything already in the `Structure` collection. This has been verified at the two places where it's called from the two overloads of `create`.