Add method `TileMap::area()`
This feels maybe a bit silly, since the origin is always `{0, 0}`, so all we really need is the `size()`. Though there are methods on `Rectangle` that are useful, and not available when working strictly with `Vector`.

It should be noted that although a new `include` was added, the `Rectangle` header was already transitively included by `PointInRectangleRange`, so shouldn't add to the compile time.