Update `isInCcRange` to use `commRange`
Note that `commRange` is affected by the operational status of the structure. The previous code allowed for `commRange` effects from a `CommandCenter` that was still under construction. Since early versions of the game, the `SeedLander` has been updated to provide `commRange`, which means we no longer need the special case hack that provides `commRange` from a `CommandCenter` that is still under construction.