Update parameters to `roadAnimationName`
It was a bit weird having a free function in `Road.cpp` that took a `Road` as it's first parameter. That made it seem a bit like it should be a member function. Makes sense to use the actual member function that calls `roadAnimationName` for the part that needs access to `Road`.