Update NAS2D submodule for `AnimationSequence` update
In particular, we need the destructor updates that allow an `AnimationSet` to be destructed even with an incomplete type for `AnimationFrame`. Since `AnimationSet` doesn't have an explicit destructor, it would get an implicit one in any translation unit that includes the `AnimationSet` header and needs to destruct such an object. Such an implicit destructor needs to be able to destruct all of the fields, which includes either the `std::vector<AnimationFrame>`, or the updated `AnimationSequence` field. The updated `AnimationSequence` field now has it's own explicit destructor, which isolates calling code from needing a complete type for `AnimationFrame`.

This update will be needed to rework how `Sprite` instances are created from the `AnimationSet` data loaded from `.sprite` files.