Remove outdated Xcode project file
The Xcode project file is so far out of date as to be unusable in it's current form. As we don't have the expertise to update it in the near future, perhaps we should remove it for now, so as not to mislead people into thinking we have a working Xcode project setup.

Any help from someone with access to Xcode and can figure out how to setup a working project would be appreciated.

For now MacOS (and Linux) users should use the `makefile` to do a source build.