Use explicit path types with `Filesystem` calls
These is needed for the recent update where `RealPath` and `VirtualPath` are now distinct types. The change was made so swapping the expected type would result in a compile time error, rather than just accepting the wrong path type as equivalent. As the two types are no longer equivalent, we need to be much more explicit about which type we are constructing.